feat: decouple app and server
