+++ packages/database/package.json
@@ -40,14 +40,14 @@
     "typeorm": "0.3.6"
   },
   "scripts": {
-    "": "─────────────────────────────────────────────── core utils ",
+    "h1": "─────────────────────────────────────────────── core utils ",
     "start": "node dist/server",
     "build": "rm -rf ./dist && tsc --declaration",
     "dev": "yarn build && yarn start",
     "graph": "madge --image graph.png",
     "ci": "yarn lint:no-fix && yarn build",
     "default:env": "sh setup.sh",
-    "": "──────────────────────────────────────────── test configs ",
+    "h2": "──────────────────────────────────────────── test configs ",
     "jest": "NODE_ENV=test jest",
     "tconf:ci": "yarn jest -c ./tests/configs/ci.ts",
     "tconf:pull": "yarn jest -c ./tests/configs/pull.ts",
@@ -56,7 +56,7 @@
     "tconf:all": "yarn jest -c ./tests/configs/all.ts",
     "tconf:server": "yarn jest -c ./tests/configs/server.ts",
     "tconf:group": "yarn jest -c ./tests/configs/group.ts",
-    "": "───────────────────────────────────────────── test aliases ",
+    "h3": "───────────────────────────────────────────── test aliases ",
     "test:ci": "yarn tconf:ci",
     "test:civ": "yarn test:ci --silent=false",
     "test:cid": "yarn test:civ --detectOpenHandles",
@@ -73,10 +73,10 @@
     "test:g": "yarn tconf:group graph",
     "test:u": "yarn tconf:group user",
     "test:b": "yarn tconf:group --silent=false base",
-    "": "────────────────────────────────────────────────── linting ",
+    "h4": "────────────────────────────────────────────────── linting ",
     "lint": "eslint --fix --ext .js,.ts .",
     "lint:no-fix": "eslint --ext .js,.ts .",
-    "": "───────────────────────────────────────────── sql commands ",
+    "h5": "───────────────────────────────────────────── sql commands ",
     "dump": "ts-node ./src/api/sql/dump.ts",
     "restore": "ts-node ./src/api/sql/restore.ts",
     "restore:file": "ts-node ./src/api/sql/restore-from-file.ts",
@@ -85,12 +85,12 @@
     "reset:mc": "ts-node ./src/api/sql/reset/moduleCondensed.ts",
     "reset:m": "ts-node ./src/api/sql/reset/module.ts",
     "reset:d": "ts-node ./src/api/sql/reset/degree.ts",
-    "": "──────────────────────────────────────────────── migration ",
+    "h6": "──────────────────────────────────────────────── migration ",
     "typeorm:ds": "typeorm-ts-node-esm -d ./src/config/index.ts",
     "mc": "yarn typeorm:ds schema:log",
     "mg": "yarn typeorm:ds migration:generate ./src/migrations/migration-name",
     "mr": "yarn typeorm:ds migration:run",
-    "": "───────────────────────────────────────────────── yarn API ",
+    "h7": "───────────────────────────────────────────────── yarn API ",
     "module:get-codes": "ts-node ./src/api/module/get-codes.ts",
     "module:get": "ts-node ./src/api/module/get.ts",
     "module:pull": "ts-node ./src/api/module/pull.ts",
